+++ src/components/layout/graph/graph-layout/GraphLayout.stories.tsx
@@ -338,6 +338,7 @@ export const WithGraphAndSettings: Story = {
     ]);
     const [setting, setSetting] = useState<GraphSideSettingValue>({
       nodeSize: 1,
+      textSize: 1,
       lineSize: 1,
       showTags: false,
       repulsion: 1500,