+++ src/components/ui/graph/graph-side/GraphSideContent.stories.tsx
@@ -37,6 +37,7 @@ const WithSearchGroupAndSetting = () => {
   ]);
   const [setting, setSetting] = useState<GraphSideSettingValue>({
     nodeSize: 1,
+    textSize: 1,
     lineSize: 1,
     showTags: false,
     repulsion: 1500,